+++ core/src/db/DB.h
@@ -104,7 +104,13 @@ class DB {
     DeleteVector(const std::string& collection_id, IDNumber vector_id) = 0;
 
     virtual Status
+    DeleteEntity(const std::string& collection_id, IDNumber entity_id) = 0;
+
+    virtual Status
     DeleteVectors(const std::string& collection_id, IDNumbers vector_ids) = 0;
+
+    virtual Status
+    DeleteEntities(const std::string& collection_id, IDNumbers entity_ids) = 0;
 
     virtual Status
     Flush(const std::string& collection_id) = 0;